requirements.txt
alembic==1.13.2
asn1crypto==1.5.1
blinker==1.8.2
Brotli==1.1.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.7
cryptography==44.0.3
cssselect2==0.8.0
cybersource-rest-client-python==0.0.64
DateTime==5.5
Deprecated==1.2.18
dnspython==2.6.1
dotenv==0.9.9
email_validator==2.2.0
Flask==3.0.3
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.57.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
limits==5.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ngrok==1.4.0
ordered-set==4.1.0
packaging==24.1
paystackapi==2.1.3
pg8000==1.31.2
pillow==11.2.1
psycopg==3.2.3
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.22.0
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
reportlab==4.4.0
requests==2.32.3
rich==13.9.4
scramp==1.4.5
six==1.16.0
SQLAlchemy==2.0.35
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.17.2
WTForms==3.1.2
zope.interface==7.2
zopfli==0.2.3.post1
